requirements_new.txt
amqp==2.1.1
beautifulsoup4==4.5.1
billiard==3.5.0.2
celery==4.0.0
click==6.6
csscompressor==0.9.4
Cython==0.25.1
defusedxml==0.4.1
Django==1.10.3
django-allauth==0.29.0
django-appconf==1.0.2
django-celery-results==1.0.1
django-cleanup==0.4.2
django-compressor==2.1
django-debug-toolbar==1.6
django-froala-editor==2.0.1
django-fs-trumbowyg==0.1.4
django-imagekit==3.3
django-taggit==0.21.3
gevent==1.1.2
greenlet==0.4.10
gunicorn==19.6.0
html5lib==0.999999999
itsdangerous==0.24
Jinja2==2.8
kombu==4.0.0
lxml==3.6.4
Markdown==2.6.7
MarkupSafe==0.23
msgpack-python==0.4.8
oauthlib==2.0.1
pilkit==1.1.13
Pillow==3.4.2
ply==3.9
PyJWT==1.4.2
python-openid==2.2.5
python-social-auth==0.2.21
python3-openid==3.0.10
pytz==2016.7
rcssmin==1.0.6
requests==2.12.2
requests-oauthlib==0.7.0
rjsmin==1.0.12
six==1.10.0
slimit==0.8.1
sorl-thumbnail==12.3
sqlparse==0.2.2
unicode-slugify==0.1.3
Unidecode==0.4.19
vine==1.1.3
webencodings==0.5
Werkzeug==0.11.11
